Move initialization of DigitalIOManager to its own class
